Enhance dot drawing behavior
Updated the `OnLButtonDown` function in `DrawView.cpp` to draw
different dot when pen size is 4. Previously it used MoveTo/LineTo,
but that drew very small rectangle. Now it manually sets the pixels
to produce a nice thick cross that looks like dot with pen size 4 in
Microsoft Paint.
I also replaced multiple GetDocument() calls in the same function
with a single one.