Add mouse capture handling in CDrawView
Added call to 'SetCapture' in 'OnLButtonDown'. Added 'OnLButtonUp'
message handler and call to 'ReleaseCapture' in it. By default, a window
stops receiving WM_MOUSEMOVE messages if the mouse moves past the edge
of the client area. This means it's not possible to draw outside of that
area. After 'SetCapture' function is called, the window will continue to
receive WM_MOUSEMOVE messages for as long as the user holds at least one
mouse button down, even if the mouse moves outside the window. That
means it is now possible to draw a line or some shape that stretches
beyond the window, and when the window is resized, that part of the
drawing becomes visible.