Fix trying to add point to nonexistent Drawable
Added flag 'isDrawing' to 'DrawView' to keep track if drawing is in
progress. What could happen previously is that if it was clicked with
left mouse button somewhere on ribbon and then dragged the mouse on
client area, with no 'Drawable' yet created, 'OnMouseMove' would try
to add point to nonexistent object and program would crash. This is
because 'Drawable' is created only in 'OnLButtonDown' and that is only
called in client area. Even in the case there already was at least one
'Drawable' in container, 'OnMouseMove' would continue adding points to
it which is also undesireable behaviour.
With this fix, flag is set to 'TRUE' with left mouse button click in
client area only and to 'FALSE' on lifting the left mouse button.
'OnMouseMove' checks the flag and does nothing if it's set to 'FALSE'.
So, if it is clicked on ribbon and then dragged on to client area,
nothing will happen.