Improve drawing performance in CDrawView with double buffering
- Added member variables to `DrawView` for memory device context and
drawing bitmap.
- Refactored `OnDraw` to use memory device context (`memDC`).
- Removed direct device context manipulation for background drawing.
- Updated `OnUpdate` to create compatible bitmap and memory DC.
- Adjusted `OnPrepareDC` for proper mapping modes based on drawing
state.
- Enhanced `OnMouseMove` and `OnLButtonDown` to utilize `memDC`.
- Added custom handling for WM_ERASEBKGND to prevent flickering.

Overall, these changes vastly improve performance in `OnDraw`.
Previously, in each `OnDraw` every `Drawable` object would be redrawn.
Now, all of the drawing is done in mouse handlers to a memory device
context. `OnDraw` is simply doing bit block transfer of stored bitmap
to a device context for drawing on a screen or printing. That also fixes
the problem of custom drawings of pen tool's dots to be incorrectly
stretched on print preview or print.