Add cursor and canvas info to status bar
Status bar now shows cursor coordinates in pane 1 and canvas size in
pane 2.

This is accomplished by introduction of new data and function members
in CDrawView and CMainFrame classes for storing and sending data from
the view to  the main frame and updating information in status bar which
is part of the main frame.