Refactor pointer casts
When switched from "use MFC in a shared DLL" to use MFC in a Static
library", DYNAMIC_DOWNCAST no longer works for CDrawApp because it
doesn't have DACLARE_DYNAMIC in .h and IMPLEMENT_DYNAMIC in .cpp.
I could have added those macros or maybe switched to dynamic_cast, but I
decided to replace it with C-style casts (Type*) instead, not only
where it caused problems, but in whole project because I don't think
it's necessary for my purposes where I'm certain of the class I'm
working with. This change simplifies the casting method, although it
removes the runtime type checking.